Add support for OmniFocus projects.
To interact with OmniFocus, I'm using a JXA script, that's executed from Ruby.

Seen from the perspective from Alfred, this could be done a bit more efficient, probably, but this way I get a central configured, managed by a set of Ruby scripts. All the Alfred-related logic is in the Ruby scripting and nowhere else.