Replace action modifiers on OmniFocus projects with action lists.
I have three problems with action modifiers (cmd, alt, opt):

1. I can't remember them.
2. They work only on the selected project (typically the top one).
3. They don't scale to more than a couple of actions.

By introducing an action list all three problems are solved. The
downside is that invoking an action takes an additional keystroke. But
that's not a real problem in my experience.

Next up: the same thing, but then for Contacts.

Once that's done, I can start thinking about adding support for project
folders on disk, to store documents and reference material.