Reverted to Alfred's Copy to Clipboard functionality.
My own scripts for pasting in the frontmost app, either by typing or by
passing it through the clipboard, weren't working well.

Typing was really slow.

Copying and pasting didn't work if there was already something in the
clipboard that wasn't text. Plus, it was slow.

So I reverted back to Alfred's built-in Copy to Clipboard feature. Not
great, since it replaces the clipboard content, but still better than
my alternatives.