First steps for the new configuration system.
The new configuration class (NewConfig) is added next to the old one.
I'll slowly move over the functionality of the old one, after which
I'll delete the old one completely and remove the NewConfig class
to Config.

In this first step the NewConfig is used only for focusing on an area.

The way this now works is that FocusPlugins are instantiated and
configured on demand, and then executed one by one. The core system
isn't aware of which plugins are available. In other words: adding a
new FocusPlugin is a very isolated tasks. That was the whole point!