Add icons to various apps for the Alfred popups.
This is not perfect, since some icons have to be "hardcoded" in Alfred
and can't be set dynamically to whatever is the user's preference for
the area of focus. At least: I have no idea on how to do that.

(Maybe I can make Alfred point to a file that is in turn a symlink to
an app icon, that I then update whenever the focus is changed? Hmm...
Something to try out once. Maybe.)