Use generic item types instead of plugin names for commands.
Using the plugin name directly - omnifocus, bear, and so on - was a bit
silly, as it would defeat the whole purpose of this tool!

The idea was to be able to do a 'vps note plain Title' to create a new
plain note, independent of the underlying note provider.

That's now fixed. And because of that the help text is also better,
because it only lists types that are actually enabled.

And, the configuration now enforces that at most one plugin is active
for a specific type.

Things are getting better and better! Except for the tests, of course...