Add lots of features to the Bear plugin.
The biggest feature: templates. The basic variable subsitutions have
been replaced with Liquid templates. It's now possible to define
templates for the title, the text and the text, all on two levels:
1) for all notes and 2) for specific note types.

Of course, I documented this in the README.

The second feature, not yet documented, is search (or better: find).
This needs more some work. I'll have this done later today, I hope.
Including the documentation of course.