Add support for iA Writer, to replace Bear.
I've used for a little under a year now, and the fact that notes
are in a database instead of plain textfiles on disk is starting to
bother me more and more. Some issues:

- Changing notes in bulk. Easy to do on the command line, hopeless in
  Bear. And also slow.
- Versioning. Bear notes have no versioning whatsoever. That's scary, especially since it autosaves.

Tags in Bear are nice, but I noticed that I started to overuse them.
After cleaning them up, very few tags are left.

That's why I've decided to move back to plaintext files. I've exported
all my notes from the Bear database with some nifty scripting for each
area, imported them in Git repositories, and updated this tool to
support them.