Added support for caching command output to speed up VPS.
This is specifically useful for the `contacts` and `groups`
plugins, as Apple Contacts can be a bit slow for large groups,
while at the same time contacts don't change that often.

Caching is **disabled** by default, to prevent surprises.

To enable, just add a `cache: true` in the configuration of the
`contacts` and/or `groups` plugins.

To flush all caches, run `vps area flush` from the command line.