Add support for "Today's note".
Currently this works only for iA Writer.

I might remove it again, if I decide a "Today note" doesn't work for me after all.

If I decide to keep it, I'll have to generalize it so that it also works for other apps, like Bear.

By the way: this update also makes sure no two notes with the same name exist, by opening the existing note instead of creating a new one.