Add support for setting the filename for iA Writer notes.
Previously I used the same text for the filename and the title of the note.
Now that I'm writing notes differently, I want to be able to influence them
separately.

For example my "Today's note" has a filename of the date in YY-MM-DD format,
whereas I like the title to be "Daily log for DD-MM-YY". This is now possible.