Add support for Obsidian.
This plugin is a quick and dirty blatant copy of the one for iA Writer.
No code reuse, just copy and paste, search and replace. Mostly.

It does however mean that all features from the iA Writer plugin also
work with Obsidian.

Since Obsidian uses plain text files, there's not a lot of scripting to
be done. Obsidian supports one URL action: `open`. Which is all that's needed!

The tricky bit was to activate Obsidian first after creating a new file, then
giving it about half a second, and only then triggering the URL action. This is
needed to have Obsidian discover the new file in time. Worst case is that
Obsidian won't find it, after which I can just run the same command again.
Nothing really breaks!