Alfred works again! That was actually easier than I thought.
Here is where the refactoring really payed of. I threw out a lot of code
and introduced a single `area commands <type> <id>` command that generates
the list of commands available to a single instance of some type. I'm
using this command all over the place in Alfred!

To really finish it up, I should now fix the 3 broken plugins, probably.