Add "yesterday" and "tomorrow" to the Today note context.
Nothing fancy yet. It doesn't know about weekends or anything. I did this only
to setup a template that has, at the top:

    < [[yesterday]] | [[tomorrow]] >

...for easier navigation from note to note.

It's error-prone, because "yesterday" might not exist (weekends), and "tomorrow"
definitely doesn't exist yet.

The first problem I could solve by actually looking for the previous today note
first. The second is harder. The neat thing to do would be to update yesterday's
note when creating the one for today. I'm not going to build all that just yet...