NoteSupport FileRepository: also write a file if it already exists with size 0.
The "create_or_find" function didn't write anything if the file already existed.

That's technically correct, but functionally it isn't, because Obsidian creates
files whenever you open an link to something that doesn't exist yet. That's also
great behavior.

In order to get the bost of both worlds, VPS overwrites an existing file it has
zero bytes in it.