Version 4.0! Breaking changes!
See the README for more information on how to migrate.

The two biggest changes:

1. The configuration file must now be in `~/.vps/config.yaml`
2. The Alfred workflow is now generated when changing focus.

Point 2 is the biggest one: I don't have to maintain the workflow
by hand anymore and it's not as complex as it was. Generating a
bit of duplication is not really a problem.

The nice thing of generating the workflow:

- No hotkeys/keywords for entities that aren't available.
- Icons for entities match the plugins.

AFAIK the generated workflow supports the same things as the version
I wrote by hand before.