Allow tags in Markdown notes to be written as YAML frontmatter.
Reason for doing this: Obsidian just introduced more support for frontmatter.

An awesome new Obisidan feature is setting up aliases. That allows one to link
to notes using their full name as well as any other name you use (like
abbreviations). It works really well!

Aliases are configured in the YAML frontmatter. As can tags. So, I decided to move
my tags over from a line of text at the bottom (with every tag prepended with '#'),
to a `tags` attribute in the frontmatter.