Allow commands that act on instances to disable themselves.
When calling "vps area commands <type> <ID>", the list of commands now only
shows commands that are able to act on the specified instance. By default a
command is always enabled.

This feature is used by the Teams and Alfred plugins, to disable themselves
if an event doesn't contain a meeting URL, or if no folder exists for the
project on disk, respectively.

In the process I refactored the contexts somewhat, so that they're easier to use
and require less code to create.