"Improve" formatting of names pulled from the Apple Calendar.
Improve is between quotes because although the result is better, the code is messier.
What Apple's Calendar presents as names, at least on a Microsoft Exchange account
(my work) is quite a mess. Anything goes, it seems.

The goal is to fix the names as much as possible to save me from doing the work of
manually fixing them (which I either forget or do incorrectly). It's probably
never going to be a 100% correct...