Add support for note indexes.
In case of plaintext notes on disk, the index command can build
a static index of all notes, including their aliases, as defined
in the notes themselves.

The `list` and `open` command use the index if it exists, for
better performance, and added functionality: notes can be found
and opened using their aliases as well.