Improve support for notes.
Paths to notes are kept relative, until the very last moment.
This saves a bit in the amount of output to Alfred.

When pasting a note, what's pasted is now a WikiLink instead
of just the name. I find that I never use the first, but do
use the second.